% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubSetFunctions.R
\name{SubSetClaims.BiRankFr}
\alias{SubSetClaims.BiRankFr}
\title{Take a subset of claims with the highest fraud scores}
\usage{
SubSetClaims.BiRankFr(
  x,
  Nodes,
  BeginDate,
  NrClaims = 10L,
  ExcludeInvestClaims = T
)
}
\arguments{
\item{x}{Object of type BiRankFr.}

\item{Nodes}{data.table with information on the nodes.}

\item{BeginDate}{Specifies from which date on you want to include claims.}

\item{NrClaims}{The number of claims}

\item{ExcludeInvestClaims}{Logical, exclude investigated claims.}
}
\value{
A list with the following components:

\item{ResultsClaims}{The dataframe with the fraud scores.}

\item{ClaimSubset}{A subset of the Nodes data.table.}
}
\description{
Take a subset of claims with the highest fraud scores
}
