% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubSetFunctions.R
\name{SubSetParties.BiRankFr}
\alias{SubSetParties.BiRankFr}
\title{Take a subset of parties with the highest scores}
\usage{
SubSetParties.BiRankFr(x, Nodes, NrParties = 10L)
}
\arguments{
\item{x}{Object of type BiRankFr.}

\item{Nodes}{data.table with information on the nodes.}

\item{NrParties}{Number of parties that you want to subset.}
}
\value{
A list with the following components:

\item{ResultsParties}{The dataframe with the scores for the parties.}

\item{PartiesSubset}{A subset of the Nodes data.table.}
}
\description{
Take a subset of parties with the highest scores
}
